Database saving, business logic and controllers are tested with swagger

+++ src/main/java/com/todeb/batuhanayyildiz/creditapplicationsystem/service/CreditLimitService.java
@@ -28,7 +28,7 @@ protected double calculateCreditLimit(int creditMultiplier,int monthlyIncome,int
         }
 
     else{
-            return monthlyIncome*creditMultiplier;
+            return (double)monthlyIncome*(double)creditMultiplier;
         }
     }
